requirements.txt
pandas
apache-airflow==2.6.1